+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_helpers.py
@@ -4,7 +4,6 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
-import re
 import six
 from azure.core.credentials import AzureKeyCredential
 from azure.core.pipeline.policies import AzureKeyCredentialPolicy
@@ -14,7 +13,6 @@ from azure.core.exceptions import (
     ResourceExistsError,
     ClientAuthenticationError
 )
-import azure.ai.formrecognizer._models as models
 
 POLLING_INTERVAL = 5
 COGNITIVE_KEY_HEADER = "Ocp-Apim-Subscription-Key"
@@ -25,15 +23,6 @@ error_map = {
     409: ResourceExistsError,
     401: ClientAuthenticationError
 }
-
-
-def get_bounding_box(field):
-    return [
-        models.Point(x=field.bounding_box[0], y=field.bounding_box[1]),
-        models.Point(x=field.bounding_box[2], y=field.bounding_box[3]),
-        models.Point(x=field.bounding_box[4], y=field.bounding_box[5]),
-        models.Point(x=field.bounding_box[6], y=field.bounding_box[7])
-    ] if field.bounding_box else None
 
 
 def adjust_value_type(value_type):
@@ -60,53 +49,6 @@ def adjust_text_angle(text_angle):
     if text_angle > 180:
         text_angle -= 360
     return text_angle
-
-
-def get_elements(field, read_result):
-    text_elements = []
-
-    for item in field.elements:
-        nums = [int(s) for s in re.findall(r"\d+", item)]
-        read = nums[0]
-        line = nums[1]
-        if len(nums) == 3:
-            word = nums[2]
-            ocr_word = read_result[read].lines[line].words[word]
-            extracted_word = models.FormWord._from_generated(ocr_word, page=read+1)  # pylint: disable=protected-access
-            text_elements.append(extracted_word)
-            continue
-        ocr_line = read_result[read].lines[line]
-        extracted_line = models.FormLine._from_generated(ocr_line, page=read+1)  # pylint: disable=protected-access
-        text_elements.append(extracted_line)
-    return text_elements
-
-
-def get_field_value(field, value, read_result):  # pylint: disable=too-many-return-statements
-    if value is None:
-        return value
-    if value.type == "string":
-        return value.value_string
-    if value.type == "number":
-        return value.value_number
-    if value.type == "integer":
-        return value.value_integer
-    if value.type == "date":
-        return value.value_date
-    if value.type == "phoneNumber":
-        return value.value_phone_number
-    if value.type == "time":
-        return value.value_time
-    if value.type == "array":
-        return [
-            models.FormField._from_generated(field, value, read_result)  # pylint: disable=protected-access
-            for value in value.value_array
-        ]
-    if value.type == "object":
-        return {
-            key: models.FormField._from_generated(key, value, read_result)  # pylint: disable=protected-access
-            for key, value in value.value_object.items()
-        }
-    return None
 
 
 def get_authentication_policy(credential):
