+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_models.py
@@ -6,16 +6,70 @@
 
 # pylint: disable=protected-access
 
+import re
 from enum import Enum
 from collections import namedtuple
 from ._helpers import (
-    get_field_value,
-    get_bounding_box,
-    get_elements,
     adjust_value_type,
     adjust_text_angle,
     adjust_confidence
 )
+
+
+def get_bounding_box(field):
+    return [
+        Point(x=field.bounding_box[0], y=field.bounding_box[1]),
+        Point(x=field.bounding_box[2], y=field.bounding_box[3]),
+        Point(x=field.bounding_box[4], y=field.bounding_box[5]),
+        Point(x=field.bounding_box[6], y=field.bounding_box[7])
+    ] if field.bounding_box else None
+
+
+def get_elements(field, read_result):
+    text_elements = []
+
+    for item in field.elements:
+        nums = [int(s) for s in re.findall(r"\d+", item)]
+        read = nums[0]
+        line = nums[1]
+        if len(nums) == 3:
+            word = nums[2]
+            ocr_word = read_result[read].lines[line].words[word]
+            extracted_word = FormWord._from_generated(ocr_word, page=read+1)  # pylint: disable=protected-access
+            text_elements.append(extracted_word)
+            continue
+        ocr_line = read_result[read].lines[line]
+        extracted_line = FormLine._from_generated(ocr_line, page=read+1)  # pylint: disable=protected-access
+        text_elements.append(extracted_line)
+    return text_elements
+
+
+def get_field_value(field, value, read_result):  # pylint: disable=too-many-return-statements
+    if value is None:
+        return value
+    if value.type == "string":
+        return value.value_string
+    if value.type == "number":
+        return value.value_number
+    if value.type == "integer":
+        return value.value_integer
+    if value.type == "date":
+        return value.value_date
+    if value.type == "phoneNumber":
+        return value.value_phone_number
+    if value.type == "time":
+        return value.value_time
+    if value.type == "array":
+        return [
+            FormField._from_generated(field, value, read_result)  # pylint: disable=protected-access
+            for value in value.value_array
+        ]
+    if value.type == "object":
+        return {
+            key: FormField._from_generated(key, value, read_result)  # pylint: disable=protected-access
+            for key, value in value.value_object.items()
+        }
+    return None
 
 
 class FieldValueType(str, Enum):
